requirements.txt
Django==4.1.7
psycopg2==2.9.5
graphene_django==3.0.0
django-graphql-jwt==0.3.4
django-cors-headers==3.14.0
